section 9 almost ended

+++ db/db_user.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm.session import Session
 from db.hash import Hash
 from db.models import DbUser
@@ -21,13 +22,25 @@ def get_all_users(db: Session):
 
 #db get user by id
 def get_user(db: Session, id: int):
-    #Handle any exceptions
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
+    return user
+
+def get_user_by_username(db: Session, username : str):
+    user = db.query(DbUser).filter(DbUser.username == username).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {username} not found')
+    return user
 
 #db update user
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
-    #Handle any exceptions
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -39,7 +52,9 @@ def update_user(db: Session, id: int, request: UserBase):
 #db delete user
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
-    #Handle any exceptions
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id} not found')
     db.delete(user)
     db.commit()
     return 'ok'
